+++ bin/circleci-rubocop-pr
@@ -1,0 +1,5 @@
+#!/usr/bin/env ruby
+
+require "circleci/rubocop/pr"
+
+Circleci::Rubocop::Pr.create_if_needed(git_username: ARGV.shift, git_email: ARGV.shift, git_branches: ARGV.empty? ? ["master"] : ARGV)
