--- lib/circleci/bundle/update/pr/version.rb
@@ -1,9 +1,0 @@
-module Circleci
-  module Bundle
-    module Update
-      module Pr
-        VERSION = "1.3.5"
-      end
-    end
-  end
-end
