+++ lib/circleci/rubocop/pr/version.rb
@@ -1,0 +1,7 @@
+module Circleci
+  module Rubocop
+    module Pr
+      VERSION = "0.0.0"
+    end
+  end
+end
